Move DisplayForm component into Form folder
