+++ src/pages/Regulatory/CookiePolicy.jsx
@@ -1,6 +1,6 @@
 import { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
-import DisplayForm from '../../components/DisplayForm';
+import DisplayForm from '../../components/Forms/DisplayForm';
 import {
   FIELD_RADIO,
   CHECKED_TRUE,
