+++ src/pages/SignIn/ResendRequestPasswordReset.jsx
@@ -16,7 +16,7 @@ import {
   REQUEST_PASSWORD_RESET_CONFIRMATION_URL,
   REQUEST_PASSWORD_RESET_URL,
 } from '../../constants/AppUrlConstants';
-import DisplayForm from '../../components/DisplayForm';
+import DisplayForm from '../../components/Forms/DisplayForm';
 
 const SupportingText = () => (
   <p className="govuk-body">Emails sometimes take a few minutes to arrive. If you did not receive the link, you can request a new one.</p>
